refactor: extract portal_theme http feature into ferriskey-api-portal-theme
