+++ libs/ferriskey-api-portal-theme/src/handlers/activate_theme.rs
@@ -4,16 +4,13 @@ use axum::{
 };
 use ferriskey_core::domain::{
     authentication::value_objects::Identity,
-    portal_theme::{
-        entities::PortalTheme,
-        ports::{GetThemeByIdInput, PortalThemeService},
-    },
+    portal_theme::ports::{GetThemeByIdInput, PortalThemeService},
 };
 use serde::{Deserialize, Serialize};
 use utoipa::ToSchema;
 use uuid::Uuid;
 
-use crate::application::http::server::{
+use ferriskey_api_core::{
     api_entities::{
         api_error::{ApiError, ApiErrorResponse},
         response::Response,
@@ -21,37 +18,37 @@ use crate::application::http::server::{
     app_state::AppState,
 };
 
-#[derive(Debug, Serialize, Deserialize, ToSchema)]
-pub struct GetThemeByIdResponse {
-    pub data: PortalTheme,
+#[derive(Debug, Clone, Serialize, Deserialize, ToSchema)]
+pub struct ActivateThemeResponse {
+    pub message: String,
 }
 
 #[utoipa::path(
-    get,
-    path = "/portal/themes/{theme_id}",
+    post,
+    path = "/portal/themes/{theme_id}/activate",
     tag = "portal-theme",
-    summary = "Get a portal theme",
-    description = "Retrieves a single portal theme, including design tokens, layout reference, and all seven page trees.",
+    summary = "Activate a portal theme",
+    description = "Marks the given theme as the realm's active theme. Stored in realm_settings.portal_theme_id.",
     params(
         ("realm_name" = String, Path, description = "Name of the realm"),
         ("theme_id" = Uuid, Path, description = "Portal theme ID"),
     ),
     responses(
-        (status = 200, description = "Theme retrieved successfully", body = GetThemeByIdResponse),
+        (status = 200, description = "Theme activated successfully", body = ActivateThemeResponse),
         (status = 404, description = "Theme not found", body = ApiErrorResponse),
         (status = 401, description = "Unauthorized", body = ApiErrorResponse),
         (status = 403, description = "Insufficient permissions", body = ApiErrorResponse),
         (status = 500, description = "Internal server error", body = ApiErrorResponse),
     ),
 )]
-pub async fn get_theme_by_id(
+pub async fn activate_theme(
     Path((realm_name, theme_id)): Path<(String, Uuid)>,
     State(state): State<AppState>,
     Extension(identity): Extension<Identity>,
-) -> Result<Response<GetThemeByIdResponse>, ApiError> {
-    let theme = state
+) -> Result<Response<ActivateThemeResponse>, ApiError> {
+    state
         .service
-        .get_theme_by_id(
+        .activate_theme(
             identity,
             GetThemeByIdInput {
                 realm_name,
@@ -61,5 +58,7 @@ pub async fn get_theme_by_id(
         .await
         .map_err(ApiError::from)?;
 
-    Ok(Response::OK(GetThemeByIdResponse { data: theme }))
+    Ok(Response::OK(ActivateThemeResponse {
+        message: "Portal theme activated successfully".to_string(),
+    }))
 }
