+++ libs/ferriskey-api-portal-theme/src/handlers/get_theme.rs
@@ -5,14 +5,14 @@ use axum::{
 use ferriskey_core::domain::{
     authentication::value_objects::Identity,
     portal_theme::{
-        entities::PortalTheme,
-        ports::{ListThemesInput, PortalThemeService},
+        entities::PortalThemeConfig,
+        ports::{GetThemeInput, PortalThemeService},
     },
 };
 use serde::{Deserialize, Serialize};
 use utoipa::ToSchema;
 
-use crate::application::http::server::{
+use ferriskey_api_core::{
     api_entities::{
         api_error::{ApiError, ApiErrorResponse},
         response::Response,
@@ -20,37 +20,37 @@ use crate::application::http::server::{
     app_state::AppState,
 };
 
-#[derive(Debug, Serialize, Deserialize, ToSchema)]
-pub struct ListThemesResponse {
-    pub data: Vec<PortalTheme>,
+#[derive(Debug, Serialize, Deserialize, ToSchema, PartialEq)]
+pub struct GetThemeResponse {
+    pub data: PortalThemeConfig,
 }
 
 #[utoipa::path(
     get,
-    path = "/portal/themes",
+    path = "/portal/theme",
     tag = "portal-theme",
-    summary = "List portal themes",
-    description = "Lists every portal theme defined in the realm. Requires manage_realm permission.",
+    summary = "Get portal theme",
+    description = "Retrieves the portal theme configuration for the realm. Requires manage_realm permission. Falls back to defaults when no configuration is stored.",
     params(
         ("realm_name" = String, Path, description = "Name of the realm"),
     ),
     responses(
-        (status = 200, description = "Themes retrieved successfully", body = ListThemesResponse),
+        (status = 200, description = "Theme configuration retrieved successfully", body = GetThemeResponse),
         (status = 401, description = "Unauthorized", body = ApiErrorResponse),
         (status = 403, description = "Insufficient permissions", body = ApiErrorResponse),
         (status = 500, description = "Internal server error", body = ApiErrorResponse),
     ),
 )]
-pub async fn list_themes(
+pub async fn get_theme(
     Path(realm_name): Path<String>,
     State(state): State<AppState>,
     Extension(identity): Extension<Identity>,
-) -> Result<Response<ListThemesResponse>, ApiError> {
-    let themes = state
+) -> Result<Response<GetThemeResponse>, ApiError> {
+    let config = state
         .service
-        .list_themes(identity, ListThemesInput { realm_name })
+        .get_theme(identity, GetThemeInput { realm_name })
         .await
         .map_err(ApiError::from)?;
 
-    Ok(Response::OK(ListThemesResponse { data: themes }))
+    Ok(Response::OK(GetThemeResponse { data: config }))
 }
