+++ libs/ferriskey-api-portal-theme/src/handlers/update_theme.rs
@@ -6,65 +6,61 @@ use ferriskey_core::domain::{
     authentication::value_objects::Identity,
     portal_theme::{
         entities::PortalTheme,
-        ports::{CreateThemeInput, PortalThemeService},
+        ports::{PortalThemeService, UpdateThemeInput},
     },
 };
 use serde::{Deserialize, Serialize};
 use utoipa::ToSchema;
 
-use crate::application::http::{
-    portal_theme::validators::CreateThemeValidator,
-    server::{
-        api_entities::{
-            api_error::{ApiError, ApiErrorResponse, ValidateJson},
-            response::Response,
-        },
-        app_state::AppState,
+use crate::validators::UpdateThemeValidator;
+use ferriskey_api_core::{
+    api_entities::{
+        api_error::{ApiError, ApiErrorResponse, ValidateJson},
+        response::Response,
     },
+    app_state::AppState,
 };
 
-#[derive(Debug, Serialize, Deserialize, ToSchema)]
-pub struct CreateThemeResponse {
+#[derive(Debug, Serialize, Deserialize, ToSchema, PartialEq)]
+pub struct UpdateThemeResponse {
     pub data: PortalTheme,
 }
 
 #[utoipa::path(
-    post,
-    path = "/portal/themes",
+    put,
+    path = "/portal/theme",
     tag = "portal-theme",
-    summary = "Create a portal theme",
-    description = "Creates a new portal theme in the realm with empty page trees. Requires manage_realm permission.",
+    summary = "Update portal theme",
+    description = "Upserts the portal theme configuration for the realm. Requires manage_realm permission.",
     params(
         ("realm_name" = String, Path, description = "Name of the realm"),
     ),
-    request_body = CreateThemeValidator,
+    request_body = UpdateThemeValidator,
     responses(
-        (status = 201, description = "Theme created successfully", body = CreateThemeResponse),
+        (status = 200, description = "Theme configuration updated successfully", body = UpdateThemeResponse),
         (status = 400, description = "Invalid request data", body = ApiErrorResponse),
         (status = 401, description = "Unauthorized", body = ApiErrorResponse),
         (status = 403, description = "Insufficient permissions", body = ApiErrorResponse),
         (status = 500, description = "Internal server error", body = ApiErrorResponse),
     ),
 )]
-pub async fn create_theme(
+pub async fn update_theme(
     Path(realm_name): Path<String>,
     State(state): State<AppState>,
     Extension(identity): Extension<Identity>,
-    ValidateJson(payload): ValidateJson<CreateThemeValidator>,
-) -> Result<Response<CreateThemeResponse>, ApiError> {
+    ValidateJson(payload): ValidateJson<UpdateThemeValidator>,
+) -> Result<Response<UpdateThemeResponse>, ApiError> {
     let theme = state
         .service
-        .create_theme(
+        .update_theme(
             identity,
-            CreateThemeInput {
+            UpdateThemeInput {
                 realm_name,
-                name: payload.name,
-                layout_id: payload.layout_id,
                 config: payload.config,
             },
         )
         .await
         .map_err(ApiError::from)?;
 
-    Ok(Response::Created(CreateThemeResponse { data: theme }))
+    Ok(Response::Updated(UpdateThemeResponse { data: theme }))
 }
